Refactor
Use kb_ naming convention for my files

Move ble hid into its own file

Delete unused code from main
