+++ 01_FW/src/utils.h
@@ -1,5 +1,6 @@
-#ifndef MY_UTILS
-#define MY_UTILS
+#ifndef KB_UTILS
+#define KB_UTILS
+
 #define CIRCULAR_INCREMENT(i, min, max) do{if(i < max) {i++;} else {i = min;}}while(0)
 
 #endif
